+++ src/core/index.ts
@@ -8,3 +8,4 @@ export * from './value-diff/index.js';
 export * from './schema-diff/index.js';
 export * from './schema-serializer/index.js';
 export * from './schema-patch/index.js';
+export * from './schema-formula/index.js';
