+++ src/core/schema-formula/index.ts
@@ -1,0 +1,4 @@
+export type { FormulaDependency } from './core/index.js';
+export { ResolvedDependency } from './core/index.js';
+export { FormulaSerializer } from './serialization/index.js';
+export { FormulaPathBuilder } from './serialization/index.js';
